Move inputs
